--- client/src/components/Header/index.js
@@ -1,24 +1,0 @@
-import React from 'react';
-
-function Header() {
-    return (
-        <div id="header">
-            <Title />
-            <UserBtn />
-        </div>
-    )
-}
-
-function Title() {
-    return (
-        <div id="title" >Compass</div>
-    )
-}
-
-function UserBtn() {
-    return (
-        <div id="userBtn" >Login</div>
-    )
-}
-
-export default Header;
